backend/requirements.txt
Django>=5.1
djangorestframework>=3.14
django-cors-headers>=4.3
openai>=1.0
python-dotenv>=1.0
